# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FinFlow API",
    "version": "1.0.0",
    "description": "All endpoints require an authenticated session (Flask-Login cookie, set by /login) and are scoped to the logged-in user's own accounts. State-changing requests (POST/PATCH/DELETE) also require a CSRF token, sent as header `X-CSRFToken` (get one from the `csrf-token` <meta> tag on any page, or the `csrf_token` hidden field on the login/signup forms)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "X-CSRFToken",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Required on POST/PATCH/DELETE."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/goals": {
      "get": {
        "summary": "List savings goals with on/off-track status",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create a savings goal",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "target_amount",
                  "deadline"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "target_amount": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date",
                    "description": "Must be in the future."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Invalid goal payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/goals/{goal_id}": {
      "delete": {
        "summary": "Delete a savings goal",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          },
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/reports/monthly": {
      "get": {
        "summary": "Plain-language monthly summary",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$"
            },
            "description": "Defaults to the current month."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid month format"
          }
        }
      }
    },
    "/reports/monthly/export.csv": {
      "get": {
        "summary": "Export that month's transactions as CSV",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file",
            "content": {
              "text/csv": {}
            }
          }
        }
      }
    },
    "/reports/monthly/export.pdf": {
      "get": {
        "summary": "Export that month's report as a PDF",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file",
            "content": {
              "application/pdf": {}
            }
          }
        }
      }
    },
...
  }
}